# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer authentication": []
    }
  ],
  "paths": {
    "/api/v1/airports/{id}": {
      "put": {
        "tags": [
          "Airport"
        ],
        "operationId": "updateAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAirportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Airport"
        ],
        "operationId": "switchAirportStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/airports/add": {
      "post": {
        "tags": [
          "Airport"
        ],
        "operationId": "addAirport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAirportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateAirportDto": {
        "required": [
          "IATACode",
          "city",
          "country",
          "name"
        ],
        "type": "object",
        "properties": {
          "IATACode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "Airport": {
        "required": [
          "IATACode",
          "city",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "IATACode": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer authentication": {
        "type": "http",
        "description": "JWT authentication",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}